fix: refator foc-auth
