+++ packages/plugin-foc-airdrop/src/types/types.ts
@@ -1,0 +1,10 @@
+export interface AirdropWallet {
+    userId: string | null,
+    address: string | null,
+}
+
+export interface Airdrop {
+    userId: string,
+    amount: string,
+    wallet: string,
+}
