+++ packages/plugin-foc-auth/src/index.ts
@@ -1,8 +1,8 @@
 import { Plugin } from "@elizaos/core";
 import { identityAuthAction } from "./actions/identityAuth.ts";
 import { identityQueryAction } from "./actions/identityQuery.ts";
-export { identityAuthProvider } from "./providers/identityAuth.ts";
 export { FocAuthKey } from "./constants";
+export { getUserIdFromState } from "./types/types.ts";
 
 export const focAuthPlugin: Plugin = {
     name: "foc auth plugin",
