+++ packages/plugin-smart-action/package.json
@@ -6,6 +6,7 @@
     "types": "dist/index.d.ts",
     "dependencies": {
         "@elizaos/core": "workspace:*",
+        "@elizaos/plugin-tee": "workspace:*",
         "tsup": "8.3.5",
         "vitest": "2.1.8"
     },
